Create a coordinator for the products tab split view with the same UX.
